# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/books/create":{"post":{"summary":"Create Books","operationId":"create_books_books_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Create Books Books Create Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/books/get":{"get":{"summary":"Get Books","operationId":"get_books_books_get_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"type":"object"},"type":"array","title":"Response Get Books Books Get Get"}}}}}}},"/books/googleapi":{"get":{"summary":"Get Google Books","operationId":"get_google_books_books_googleapi_get","parameters":[{"description":"Book title","required":true,"schema":{"type":"string","title":"Query","description":"Book title"},"name":"query","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/books/googleapi/freebooks":{"get":{"summary":"Get Free Google Books","operationId":"get_free_google_books_books_googleapi_freebooks_get","parameters":[{"description":"Book title","required":true,"schema":{"type":"string","title":"Query","description":"Book title"},"name":"query","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/books/add_google_book":{"post":{"summary":"Add Google Book","operationId":"add_google_book_books_add_google_book_post","parameters":[{"description":"Book id from Google API","required":true,"schema":{"type":"string","title":"Book Id","description":"Book id from Google API"},"name":"book_id","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Add Google Book Books Add Google Book Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/books/update/{book_id}":{"put":{"summary":"Update Book","operationId":"update_book_books_update__book_id__put","parameters":[{"required":true,"schema":{"type":"string","title":"Book Id"},"name":"book_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Update Book Books Update  Book Id  Put"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/books/delete/{book_id}":{"delete":{"summary":"Delete Book","operationId":"delete_book_books_delete__book_id__delete","parameters":[{"required":true,"schema":{"type":"string","title":"Book Id"},"name":"book_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Delete Book Books Delete  Book Id  Delete"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/books/deleteall":{"delete":{"summary":"Delete All Books","operationId":"delete_all_books_books_deleteall_delete","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Delete All Books Books Deleteall Delete"}}}}}}}},"components":{"schemas":{"Book":{"properties":{"google_book_id":{"type":"string","minLength":1,"title":"Google Book Id"},"title":{"type":"string","title":"Title"},"authors":{"items":{"type":"string"},"type":"array","minItems":1,"title":"Authors"},"published_date":{"type":"string","title":"Published Date"},"description":{"type":"string","title":"Description"}},"type":"object","required":["google_book_id","title","authors","description"],"title":"Book"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}